+++ NINA.Plugin.TargetScheduler.Test/Database/Schema/ProfilePreferenceTest.cs
@@ -16,6 +16,7 @@ namespace NINA.Plugin.TargetScheduler.Test.Database.Schema {
             sut.ParkOnWait.Should().BeFalse();
             sut.ExposureThrottle.Should().BeApproximately(125, 0.001);
             sut.EnableSmartPlanWindow.Should().BeTrue();
+            sut.EnablePlannerReports.Should().BeFalse();
             sut.EnableDeleteAcquiredImagesWithTarget.Should().BeTrue();
             sut.EnableSlewCenter.Should().BeTrue();
             sut.EnableStopOnHumidity.Should().BeTrue();